createTables: [refactor] separate DataLoader, TableCreator
